+++ CHANGES.md
@@ -68,6 +68,11 @@
 
     - This version has some improved checks for syntax errors.
 
+    - Fixed issue with --dump-mixed-call-parens. A trailing statement modifier
+      such as the following 'if' was incorrectly being counted as having parens:
+
+	return $class if ($old_quote - $new_quote) == 0;
+
 ## 2025 09 12
 
     - Fix git #191, where a new check to escape whitespace in a qw() quote